requirements.txt
argparse==1.2.1
distribute==0.6.24
django-tastypie==0.9.11
mimeparse==0.1.3
python-dateutil==1.5
wsgiref==0.1.2
newrelic
